forecasting-service/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0

requests==2.31.0
python-consul2==0.1.5

pandas==2.2.2
numpy<2.0 
prophet==1.1.5

psycopg2-binary==2.9.9
